nvme: add an OpenChannel 2.0 NVMe device (ocssd)
This adds a new 'ocssd' device that emulates an OpenChannel 2.0 device.

The ocssd device is an individual device but shares a lot of code with
the nvme device. Thus, some core functionality of nvme/nvme.c has been
exported for use by nvme/ocssd.c.

Thank you to the following people for their contributions to the
original qemu-nvme (github.com/OpenChannelSSD/qemu-nvme) implementation.

  Matias Bjørling <mb@lightnvm.io>
  Javier González <javier@javigon.com>
  Simon Andreas Frimann Lund <ocssd@safl.dk>
  Hans Holmberg <hans@owltronix.com>
  Jesper Devantier <contact@pseudonymous.me>
  Young Tack Jin <youngtack.jin@circuitblvd.com>

Signed-off-by: Klaus Birkelund Jensen <klaus.jensen@cnexlabs.com>